Feature/questions

+++ frontend/components/track_show/second_column/annotation_show.jsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import AnnotationBody from './annotation_body';
+import { Link } from 'react-router-dom';
 import * as AnnotateUtil from '../../../util/annotate_util';
 import * as UpvoteUtil from '../../../util/upvote_util';
 
@@ -36,9 +38,11 @@ class AnnotationShow extends React.Component {
   }
 
   deleteButton() {
-    const annotationId = this.props.match.params.annotationId;
-    this.props.deleteAnnotation(annotationId)
-      .then(() => this.props.history.push(`/tracks/${this.props.match.params.trackId}`));
+    if (confirm("Delete this annotation?")) {
+      const annotationId = this.props.match.params.annotationId;
+      this.props.deleteAnnotation(annotationId)
+        .then(() => this.props.history.push(`/tracks/${this.props.match.params.trackId}`));
+    }
   }
 
   upvote(e) {
@@ -85,16 +89,25 @@ class AnnotationShow extends React.Component {
     return show ? 
       (<div className="annotation-show-container annotation-show" style={style}>
         <h3 className="annotation-show">Prodigy Annotation</h3>
-        <p className="annotation-show">{currentAnnotation.body}</p>
-        {this.props.currentUserId === currentAnnotation.author_id && 
-          <button 
-            className="annotation-delete annotation-show" 
-            onClick={this.deleteButton}
-          >Delete</button>}
+        <div className="annotation-show"><AnnotationBody body={currentAnnotation.body}/></div>
         <div className="annotation-show-bottom annotation-show">
-          <a className="upvote-link" onClick={leftCb}><i className={`far fa-thumbs-up ${leftClass} annotation-show`} /></a>
-          <span className="upvote-count annotation-show">{sign}{upvoteCount}</span>
-          <a className="upvote-link" onClick={rightCb}><i className={`far fa-thumbs-down fa-flip-horizontal ${rightClass} annotation-show`} /></a>
+          <div className="annotation-show-upvotes annotation-show">
+            <a className="upvote-link" onClick={leftCb}><i className={`far fa-thumbs-up ${leftClass} annotation-show`} /></a>
+            <span className="upvote-count annotation-show">{sign}{upvoteCount}</span>
+            <a className="upvote-link" onClick={rightCb}><i className={`far fa-thumbs-down fa-flip-horizontal ${rightClass} annotation-show`} /></a>
+          </div>
+          <div className="annotation-show-buttons annotation-show">
+            {this.props.currentUserId === currentAnnotation.author_id &&
+              <button
+                className="annotation-delete annotation-show"
+                onClick={this.deleteButton}
+              >Delete</button>}
+            {this.props.currentUserId === currentAnnotation.author_id &&
+              <Link
+                className="annotation-edit annotation-show"
+                to={`/tracks/${currentAnnotation.track_id}/${currentAnnotation.id}/edit`}
+              >Edit</Link>}
+          </div>
         </div>
       </div>
       ) : (<div></div>)
